add config.toml

+++ main.go
@@ -6,7 +6,8 @@ import (
 	"strconv"
 	"time"
 	"flag"
-
+	"log"
+	"github.com/BurntSushi/toml"
 	"github.com/nlopes/slack"
 )
 
@@ -20,14 +21,39 @@ func ctv(ts string) time.Time {
 
 func usage() {
 	fmt.Println("slack-cleaner is a tool for removing messages into a slack channel")
+	fmt.Println("")
 	fmt.Println("Usage:")
 	fmt.Println("\tslack-cleaner <options> channel")
 	fmt.Println("\tchannel\t:The name of the channel")
+	fmt.Println("")
 	fmt.Println("options:")
 
-	fmt.Printf("\t-date\t:Remove every message before that date 'YYYYMMDD'\n")
-	fmt.Printf("\t-user\t:Message user\n\n")
-	fmt.Printf("\t-bot\t:bot id\n\n")
+	fmt.Println("\t-date\t:Remove every message before that date 'YYYYMMDD'")
+	fmt.Println("\t-user\t:Message user")
+	fmt.Println("\t-bot\t:bot id")
+	fmt.Println("\t-n\t:dry run")
+	fmt.Println("")
+}
+
+// Config ...
+type Config struct {
+	SlackID   string
+}
+
+// ReadConfig ...
+func ReadConfig() Config {
+	var configfile = "config.toml"
+	_, err := os.Stat(configfile)
+	if err != nil {
+		log.Fatal("Config file is missing: ", configfile)
+	}
+
+	var config Config
+	if _, err := toml.DecodeFile(configfile, &config); err != nil {
+		log.Fatal(err)
+	}
+
+	return config
 }
 
 func main() {
@@ -36,11 +62,10 @@ func main() {
 	selectedDate := flag.String("date", "", "delete messages until this date") 
 	selectedUser := flag.String("user", "", "only for this user") 
 	selectedBotID := flag.String("bot", "", "only for this bot id") 
+	dryRun := flag.Bool("n", false, "dry run") 
 
 	flag.Parse()
 	myArgs :=  flag.Args()
-
-	fmt.Println(myArgs)
 
 	if len(myArgs) != 1 || *help {
 		usage()
@@ -49,19 +74,21 @@ func main() {
 
 	selectedChannel := myArgs[0]
 
-	fmt.Println("selectedChannel:", selectedChannel)
-	fmt.Println("date:", *selectedDate)
-	fmt.Println("user:", *selectedUser)
+	log.Println("selectedChannel:", selectedChannel)
+	log.Println("date:", *selectedDate)
+	log.Println("user:", *selectedUser)
+	log.Println("dry:", *dryRun)
 	
 	dateSelectedString := ""
 	if *selectedDate != "" {
 		const shortForm = "20060102"
 		t, _ := time.Parse(shortForm, *selectedDate)
 		dateSelectedString = strconv.FormatFloat(float64(t.Unix()), 'f', 6, 64)
-		fmt.Printf("Remove every message on <%s> before <%s>\n", selectedChannel, t.Format(shortForm))	
+		log.Printf("Remove every message on <%s> before <%s>\n", selectedChannel, t.Format(shortForm))	
 	}
 
-	api := slack.New("xoxp-11888837425-11886289616-13691440419-33c723bbbb")
+	var config = ReadConfig()
+	api := slack.New(config.SlackID)
 
 	params := &slack.HistoryParameters{
 		Latest:    dateSelectedString,
@@ -73,7 +100,7 @@ func main() {
 
 	channels, err := api.GetChannels(false)
 	if err != nil {
-		fmt.Printf("%s\n", err)
+		log.Fatal("Error :", err)
 		return
 	}
 
@@ -86,7 +113,7 @@ func main() {
 	}
 
 	if selectedChannelID == "" {
-		fmt.Printf("Channel not found <%s>\n", selectedChannel)
+		log.Fatalf("Channel not found <%s>", selectedChannel)
 		os.Exit(1)
 	}
 
@@ -95,12 +122,12 @@ func main() {
 	if *selectedUser != "" {
 		users, err := api.GetUsers()
 		if err != nil {
-			fmt.Printf("%s\n", err)
+			log.Fatal("Error :", err)
 			return		
 		}
 
 		for _, user := range users {
-			fmt.Printf("user: %s %s\n", user.ID, user.Profile.RealName)
+			log.Printf("user: %s %s\n", user.ID, user.Profile.RealName)
 			if *selectedUser == user.Profile.RealName {
 				selectedUserID = user.ID
 			}
@@ -111,14 +138,14 @@ func main() {
 
 	history, err := api.GetChannelHistory(selectedChannelID, *params)
 	if err != nil {
-		fmt.Printf("Error: %s\n", err)
+		log.Fatal("Error :", err)
 		return
 	}
 
 
 
 	for _, message := range history.Messages {
-		var skip = false
+		var skip = false;
 
 		if selectedUserID != "" && selectedUserID != message.User {
 			skip = true
@@ -129,11 +156,23 @@ func main() {
 		}		
 
 		if !skip {
-			respChannel, respTimestamp, err := api.DeleteMessage(selectedChannelID, message.Timestamp)
-			fmt.Printf("res: %s %s %s \n", respChannel, ctv(respTimestamp), err)
+
+			if (*dryRun) {
+				log.Printf("(dry) Delete : %s \n", ctv(message.Timestamp))				
+			} else {
+				log.Printf("Delete : %s %s", ctv(message.Timestamp), message.Text)
+				respChannel, respTimestamp, err := api.DeleteMessage(selectedChannelID, message.Timestamp)
+				if err != nil {
+					log.Fatal("Error :", err)
+				} else {
+					log.Printf("res: %s %s", respChannel, ctv(respTimestamp))
+				}	
+				// slack rates...
+				time.Sleep(2 * time.Second)
+			}
+		
 		}
 		
-		time.Sleep(2 * time.Second)
 	}
 
 }
